Setting up UI, preparing for presentation
I figured out how to use the Windows 8 app bar and made some good
progress on the UI by moving a bunch of the buttons to there.

Since I added a discard button to the app bar I tried to implement
functionality to delete a video and seemingly achieved it in this
commit.