Created custom popup for video details
When a user uploads a new video they can upload it with details such as
title, descriptions and tags. Further information being taken sent to
the server by the as metadata for the video is location (which will need
to be enabled) and length of video.

In this commit I started to define a custom popup where users could
enter the title and description for their video. At the bottom I placed
a Submit button, which in time I think will be the Upload button and
this dialog I defined will be the last thing the user does before
sending the video to the server.

Some bug fixes to do:
- If the video was previously recorded and I try to send the file it
doesn't work it breaks.
- There's a weird bug with the camera such that the first ~1 seconds of
the video is pretty much garbage. Look online to see if this is a common
problem and if so, what is the remedy?