Bloody massive commit!
Added so much stuff!
- Worked on login/logout behavior so I show/hide controls accordingly
- Worked on videos page. I added a bunch of dummie images and populated
a gridview with them, which is horizontally scrollable.

The prototype is looking extremely good and I'm very happy about it.

TODO:
- Create custom user control for when a user uploads a video which
contains a message notifying them that their video was sent, a specific
link to view their video as well as buttons to share the link on
facebook, twitter or youtube.